aider: Updated the path to the StyleGAN directory to be dynamically set to the 'stylegan-encoder' directory in the current working directory.
# Aider chat conversation:

USER: or was stylegan referring to the current directory in which everything takes place?

ASSISTANT: The line `sys.path.append('/path/to/stylegan')` is used to add the directory where the StyleGAN repository is located to the Python path. This allows Python to find the StyleGAN modules when they are imported in the script.

If you have a folder called `stylegan-encoder` in the current directory, that is likely the directory that contains the `dnnlib` module. You should replace `'/path/to/stylegan'` with the actual path to the `stylegan-encoder` directory on your system.

Here is the *SEARCH/REPLACE block* for this change:

StyleGAN_Encoder.py
```python
<<<<<<< SEARCH
sys.path.append('/path/to/stylegan')
=======
sys.path.append(os.path.join(os.getcwd(), 'stylegan-encoder'))
>>>>>>> REPLACE
```

This change will add the `stylegan-encoder` directory in the current working directory to the Python path.