aider: Replaced os.system with subprocess.check_call to ensure dlib is installed in the correct Python environment.
# Aider chat conversation:

USER: ~/alexhkurz-at-git/StyleGAN-Face-Morphing > python StyleGAN_Encoder.py
Requirement already satisfied: dlib in /opt/homebrew/lib/python3.11/site-packages (19.24.2)
DEPRECATION: textract 1.6.5 has a non-standard dependency specifier extract-msg<=0.29.*. pip 24.0 will enforce this behaviour change. A possible replacement is to upgrade to a newer version of textract or contact the author to suggest that they release a version with a conforming dependency specifiers. Discussion can be found at https://github.com/pypa/pip/issues/12063
Traceback (most recent call last):
  File "/Users/akurz/alexhkurz-at-git/StyleGAN-Face-Morphing/StyleGAN_Encoder.py", line 10, in <module>
    import dlib
ModuleNotFoundError: No module named 'dlib'

ASSISTANT: It seems like the `dlib` module is not being found by your Python interpreter. This could be due to a variety of reasons such as the module not being installed in the correct Python environment, or the Python interpreter not looking in the correct location for installed modules.

To ensure that `dlib` is installed in the correct Python environment, we can modify the script to use `subprocess` module instead of `os.system` for installing `dlib`. This will ensure that `dlib` is installed in the same Python environment that is running the script.

Here is the *SEARCH/REPLACE block* for this change:

StyleGAN_Encoder.py
```python
<<<<<<< SEARCH
try:
    import dlib
except ModuleNotFoundError:
    os.system('pip install dlib')
    import dlib
=======
import subprocess
try:
    import dlib
except ModuleNotFoundError:
    subprocess.check_call([sys.executable, '-m', 'pip', 'install', 'dlib'])
    import dlib
>>>>>>> REPLACE
```

This change will replace the `os.system` call with `subprocess.check_call` to ensure that `dlib` is installed in the correct Python environment.